Implemented llri::Instance (squash merge)
This subfeature introduces the Instance as the core structure of the LLRI API.
The subfeature brings the following:

1 new handle:

llri::Instance is responsible for the creation of all API objects in the application. It's the equivalent of VkInstance, DXGI_Factory and an OpenGL Context.
3 new structures:

llri::InstanceExtension Describes an optional instance extension. These extensions may or may not be available depending on the available internal API, and the system's hardware and drivers.
llri::APIValidationEXT An instance extension that enables API validation. API validation causes performance overhead but is useful in debugging scenarios.
llri::GPUValidationEXT An instance extension that enables GPU validation. GPU validation causes significant performance overhead but can be useful when debugging shaders or synchronization issues.
1 new enum type:

llri::InstanceExtensionType Describes the kind of instance extension. This value is used in InstanceExtension and is used internally to recognize the extension type and cast InstanceExtension::data to the correct underlying type
3 new functions:

llri::createInstance() Creates an instance handle with the given instance description. Passed extensions are created/enabled internally.
llri::destroyInstance() Destroys the instance and its internal extension resources. This does not include resources created through the instance, that responsibility remains up to the user.
llri::queryInstanceExtensionSupport() Queries the support of the given extension. Returns true if the extension is supported, and false if it isn't.

------------------------------
Commits:
------------------------------

commit a2c5075c356c92d7d46712f0b981c98b888a9f1f
Author: Leon Brands (Desktop) <lajbrands@outlook.com>
Date:   Mon Jul 12 22:01:54 2021 +0200

    Added a to_string function for InstanceExtensionType

commit 382e1e1cbdbdb0b05243aab826adfd3ea8a2212b
Author: Leon Brands (Desktop) <lajbrands@outlook.com>
Date:   Mon Jul 12 20:48:36 2021 +0200

    Documented instance extensions

commit 9a9cf63c783cfc1f37516c9187a7814610c24adf
Author: Leon Brands (Desktop) <lajbrands@outlook.com>
Date:   Sat Jul 10 23:42:42 2021 +0200

    llri::Instance

    Added llri::Instance:
    * Instance, InstanceDesc and createInstance()
    * InstanceExtension, InstanceExtensionType, queryInstanceExtensionSupport()
    * And current supported instance extensions: APIValidationEXT, GPUValidationEXT

    testsystem.cpp contains sample code for the instance creation API